+++ src/InputGroup/Item/Item.tsx
@@ -1,5 +1,9 @@
 import React from 'react';
-import { StyledItem, ItemProps } from './Item.styles';
+import { StyledItem } from './Item.styles';
+
+export interface ItemProps {
+  backgroundColor?: string;
+}
 
 export const Item: React.FC<ItemProps> = ({ backgroundColor, children }) => (
   <StyledItem backgroundColor={backgroundColor}>{children}</StyledItem>
